requirements.txt
databricks-sql-connector==4.0
databricks-sqlalchemy==2.0.4
databricks-vectorsearch==0.45
langchain-community==0.3.19
langchain-core==0.3.44
langchain_experimental==0.3.4
langgraph==0.3.5
langgraph-prebuilt==0.1.1
databricks-agents==0.15.0
databricks-langchain==0.2.0